LLM-summarization/requirements.txt
ffmpeg-python==0.2.0
whisper==0.0.1
openai==1.14.0
spacy==3.7.2
textgrid==1.5
python-dotenv==1.0.0
nltk==3.8.1
torch==2.2.0
tqdm==4.66.1
transformers==4.38.2
accelerate==0.28.0
scipy==1.12.0
numpy==1.26.3
scikit-learn==1.4.0
huggingface-hub==0.20.3
en_core_web_sm@ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl
